% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdayprob.R
\name{bdayprob}
\alias{bdayprob}
\title{Birthday Problem}
\usage{
bdayprob(groupsize, nsims, n = 2)
}
\description{
Function that takes the size of the group and the number of simulations to
run to calculate the probability of that group having two or more members
with the same birthday.
}
