% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FnDiff.R
\name{fnDiff}
\alias{fnDiff}
\title{fnDiff}
\usage{
fnDiff(x, n, fnctn = mean)
}
\description{
Function for testing - takes the difference in some statistics (defaults to
mean) between elements of a vector, from 1 to element n, and from n+1 to the
end of the vector.
}
