% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfrowSetup.R
\name{mfrowSetup}
\alias{mfrowSetup}
\title{mfrow Setup}
\usage{
mfrowSetup(nPlots = 1, maxCols = 1, ...)
}
\description{
Function to quickly produce mfrow setup for a number of plots and a max number
of columns you want to display Defaults to 1x1, so you can quickly use it to
reset the plot window as well.
}
